refactor: Remove legacy hierarchy system and duplicate code
- Removed unused hierarchy system components (HierarchySidebar, useHierarchy)
- Consolidated duplicate implementations (page-refactored → page, etc.)
- Removed legacy time period system
- Cleaned up unused types and state management
- Simplified type exports and removed re-exports
- Fixed TypeScript errors and improved type safety

This significantly reduces code complexity and improves maintainability
by removing ~1000 lines of unused/duplicate code.
