+++ src/frontend/nextjs-app/app/demand-planning/components/charts/BaseChart.tsx
@@ -1,27 +1,20 @@
 'use client';
 
-import { useRef, useEffect, useState } from 'react';
+import { useRef, useEffect } from 'react';
 import * as d3 from 'd3';
 
 export interface BaseChartProps {
   width?: number;
   height?: number;
-  margin?: { top: number; right: number; bottom: number; left: number };
   className?: string;
 }
 
 export default function BaseChart({
   width = 600,
   height = 400,
-  // margin is defined but used in derived charts
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  margin = { top: 20, right: 30, bottom: 40, left: 50 },
   className = ''
 }: BaseChartProps) {
   const svgRef = useRef<SVGSVGElement>(null);
-  const [tooltipVisible, setTooltipVisible] = useState(false);
-  const [tooltipContent, setTooltipContent] = useState('');
-  const [tooltipPosition, setTooltipPosition] = useState({ x: 0, y: 0 });
 
   // Create a resize observer to handle responsive behavior
   useEffect(() => {
@@ -47,19 +40,6 @@ export default function BaseChart({
     };
   }, [width, height]);
 
-  // Helper functions for tooltip - used by child components
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const showTooltip = (content: string, x: number, y: number) => {
-    setTooltipContent(content);
-    setTooltipPosition({ x, y });
-    setTooltipVisible(true);
-  };
-
-  // Helper function to hide tooltip
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const hideTooltip = () => {
-    setTooltipVisible(false);
-  };
 
   return (
     <div className={`relative ${className}`}>
@@ -72,17 +52,6 @@ export default function BaseChart({
         {/* Chart contents will be rendered by specific implementations */}
       </svg>
 
-      {/* Tooltip */}
-      {tooltipVisible && (
-        <div
-          className="absolute z-10 bg-dp-surface-secondary backdrop-blur-sm border border-dp-border-light rounded-md shadow-dp-medium pointer-events-none transform -translate-x-1/2 -translate-y-full"
-          style={{
-            left: `${tooltipPosition.x}px`,
-            top: `${tooltipPosition.y - 10}px`,
-          }}
-          dangerouslySetInnerHTML={{ __html: tooltipContent }}
-        />
-      )}
     </div>
   );
 }
