+++ src/frontend/nextjs-app/app/demand-planning/page.tsx
@@ -5,7 +5,6 @@ import DashboardLayout from './components/DashboardLayout';
 import AdjustmentPanel from './components/AdjustmentPanel';
 import AdjustmentHistoryTable from './components/AdjustmentHistoryTable';
 import { FilterSelections } from './components/FilterSidebar';
-import { HierarchySelection } from '@/app/types/demand-planning';
 import useForecast from './hooks/useForecast';
 import useAdjustmentHistory from './hooks/useAdjustmentHistory';
 import { AdjustmentData } from './components/AdjustmentModal';
@@ -24,178 +23,112 @@ export default function DemandPlanningPage() {
     dateRange: { startDate: null, endDate: null }
   });
 
-  // Keep hierarchy selections for backward compatibility with useForecast hook
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [selectedHierarchies, setSelectedHierarchies] = useState<HierarchySelection[]>([]);
-  // Initialize with all daily periods from Jan 1 to Mar 31, 2025
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [selectedTimePeriods, setSelectedTimePeriods] = useState<string[]>([]);
   const [activeTab, setActiveTab] = useState<'forecast' | 'history' | 'settings'>('forecast');
 
-  // Time periods will come from useForecast hook based on actual data
-
-  // Initialize without hardcoded selections - let users choose their own filters
+  // Initialize without hardcoded selections
   useEffect(() => {
     console.log("Page component mount - users will select their own hierarchies and filters");
-    // No hardcoded initial selections - start with empty state
-    // useForecast will use the full available date range when no periods are selected
   }, []);
 
-  // Fetch forecast data based on selections
+  // Fetch forecast data using TanStack Query
   const {
     forecastData,
     isLoading: isLoadingForecast,
     error: forecastError,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    availableDateRange,
     applyAdjustment,
-    resetAdjustments,
-    refreshForecast
+    refetch,
   } = useForecast({
-    hierarchySelections: selectedHierarchies,
-    timePeriodIds: selectedTimePeriods,
-    filterSelections
+    filterSelections,
   });
 
   // Fetch adjustment history
   const {
     adjustmentHistory,
     isLoading: isLoadingHistory,
-    // Error variable is available but not currently used in the UI
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     error: historyError,
-    refreshHistory
+    refreshHistory,
   } = useAdjustmentHistory();
 
-  // Handle filter selection changes from sidebar
-  const handleFilterSelectionChange = (selections: FilterSelections) => {
-    setFilterSelections(selections);
-    console.log("Filter selections changed:", selections);
-
-    // You can add logic here to convert filter selections to hierarchy selections
-    // if needed for backward compatibility with existing hooks
+  // Handle filter changes from FilterSidebar
+  const handleFilterChange = (newSelections: FilterSelections) => {
+    console.log("DemandPlanningPage: Filter selections changed", newSelections);
+    setFilterSelections(newSelections);
   };
 
-  // Handle tab change
-  const handleTabChange = (tab: 'forecast' | 'history' | 'settings') => {
-    setActiveTab(tab);
+  // Handle adjustment application
+  const handleApplyAdjustment = async (adjustmentData: AdjustmentData) => {
+    console.log("Applying adjustment:", adjustmentData);
+    try {
+      await applyAdjustment(adjustmentData);
+      console.log("Adjustment applied successfully");
 
-    // If switching to history tab, refresh the history data
-    if (tab === 'history') {
-      refreshHistory();
+      // Refetch adjustment history to show the new adjustment
+      await refreshHistory();
+    } catch (error) {
+      console.error("Failed to apply adjustment:", error);
     }
   };
 
-  // Handle adjustment creation
-  const handleApplyAdjustment = async (adjustment: AdjustmentData) => {
-    await applyAdjustment(adjustment);
-
-    // Refresh history after applying adjustment
-    refreshHistory();
-  };
+  // Calculate available periods based on forecast data (removed - not used)
+  // const availablePeriods = forecastData?.timePeriods || [];
 
   return (
     <DashboardLayout
       filterSelections={filterSelections}
-      onFilterSelectionChange={handleFilterSelectionChange}
+      onFilterSelectionChange={handleFilterChange}
       activeTab={activeTab}
-      onTabChange={handleTabChange}
+      onTabChange={setActiveTab}
     >
+      {/* Main content area based on active tab */}
       {activeTab === 'forecast' && (
-        <div className="grid gap-6">
-          {/* Forecast View */}
-          <div className="bg-dp-surface-primary border border-dp-border-light rounded-lg shadow-dp-light mb-6">
-            <div className="flex justify-between items-center p-5 border-b border-dp-border-light">
-              <div>
-                <h1 className="text-2xl font-medium text-dp-text-primary">Sales Forecast</h1>
-                <p className="text-dp-text-secondary mt-1">
-                  {filterSelections.inventoryItemId
-                    ? (() => {
-                        const locations = [];
-                        if (filterSelections.states.length > 0) {
-                          locations.push(`${filterSelections.states.length} state${filterSelections.states.length !== 1 ? 's' : ''}`);
-                        }
-                        if (filterSelections.dmaIds.length > 0) {
-                          locations.push(`${filterSelections.dmaIds.length} DMA${filterSelections.dmaIds.length !== 1 ? 's' : ''}`);
-                        }
-                        if (filterSelections.dcIds.length > 0) {
-                          locations.push(`${filterSelections.dcIds.length} DC${filterSelections.dcIds.length !== 1 ? 's' : ''}`);
-                        }
-                        const locationText = locations.length > 0
-                          ? `for ${locations.join(', ')}`
-                          : 'across all locations';
-                        return `Viewing forecast for Item ${filterSelections.inventoryItemId} ${locationText}`;
-                      })()
-                    : 'Loading inventory items...'
-                  }
-                </p>
+        <div className="space-y-6">
+          {/* Chart Section */}
+          <div className="bg-white dark:bg-gray-800 rounded-lg shadow p-6">
+            <Suspense fallback={
+              <div className="flex items-center justify-center h-[400px]">
+                <div className="text-gray-500">Loading charts...</div>
               </div>
-              <CacheStatus />
-            </div>
-
-            {isLoadingForecast ? (
-              <div className="bg-dp-background-tertiary rounded-lg p-4 h-80 flex items-center justify-center">
-                <div className="flex flex-col items-center">
-                  <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-dp-cfa-red mb-4"></div>
-                  <p className="text-dp-text-tertiary">Loading forecast data...</p>
+            }>
+              {isLoadingForecast ? (
+                <div className="flex items-center justify-center h-[400px]">
+                  <div className="text-gray-500">Loading forecast data...</div>
                 </div>
-              </div>
-            ) : forecastError ? (
-              <div className="bg-red-50 text-red-700 rounded-lg p-4 flex items-center">
-                <svg className="h-6 w-6 mr-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />
-                </svg>
-                <span>{forecastError}</span>
-              </div>
-            ) : forecastData ? (
-              <div className="p-4">
-                <Suspense fallback={
-                  <div className="bg-dp-background-tertiary rounded-lg p-4 h-80 flex items-center justify-center">
-                    <div className="flex flex-col items-center">
-                      <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-dp-cfa-red mb-4"></div>
-                      <p className="text-dp-text-tertiary">Loading chart...</p>
-                    </div>
-                  </div>
-                }>
-                  <ForecastCharts
-                    forecastData={forecastData}
-                  />
-                </Suspense>
-              </div>
-            ) : (
-              <div className="bg-dp-background-tertiary rounded-lg p-4 h-80 flex items-center justify-center">
-                <div className="text-center">
-                  <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-dp-cfa-red mb-4 mx-auto"></div>
-                  <p className="text-dp-text-tertiary">Preparing forecast data...</p>
-                  <p className="text-dp-text-tertiary text-sm mt-2">
-                    {filterSelections.states.length === 0 && filterSelections.dmaIds.length === 0 && filterSelections.dcIds.length === 0
-                      ? "Aggregating data across all locations..."
-                      : "Loading forecast for selected locations..."}
-                  </p>
+              ) : forecastError ? (
+                <div className="flex items-center justify-center h-[400px]">
+                  <div className="text-red-500">{forecastError}</div>
                 </div>
-              </div>
-            )}
+              ) : forecastData ? (
+                <ForecastCharts forecastData={forecastData} />
+              ) : (
+                <div className="flex items-center justify-center h-[400px]">
+                  <div className="text-gray-500">Select filters to view forecast data</div>
+                </div>
+              )}
+            </Suspense>
           </div>
 
           {/* Adjustment Panel */}
           {forecastData && (
-            <div className="max-w-md">
-              <AdjustmentPanel
-                forecastData={forecastData}
-                isLoading={isLoadingForecast}
-                onApplyAdjustment={handleApplyAdjustment}
-                onResetAdjustments={resetAdjustments}
-                onRefreshForecast={refreshForecast}
-                selectedHierarchies={selectedHierarchies}
-              />
-            </div>
+            <AdjustmentPanel
+              forecastData={forecastData}
+              isLoading={isLoadingForecast}
+              onApplyAdjustment={handleApplyAdjustment}
+              onResetAdjustments={async () => {
+                console.log("Reset adjustments not implemented");
+              }}
+              onRefreshForecast={async () => {
+                await refetch();
+              }}
+            />
           )}
         </div>
       )}
 
       {activeTab === 'history' && (
-        <div className="grid gap-6">
-          {/* History View */}
+        <div className="bg-white dark:bg-gray-800 rounded-lg shadow">
+          {historyError && (
+            <div className="p-4 text-red-500">{historyError}</div>
+          )}
           <AdjustmentHistoryTable
             entries={adjustmentHistory}
             isLoading={isLoadingHistory}
@@ -204,11 +137,9 @@ export default function DemandPlanningPage() {
       )}
 
       {activeTab === 'settings' && (
-        <div className="dp-card p-6">
-          <h1 className="text-2xl font-light mb-4 text-dp-text-primary">Settings</h1>
-          <p className="text-dp-text-secondary">
-            Dashboard settings will be implemented in a future update.
-          </p>
+        <div className="bg-white dark:bg-gray-800 rounded-lg shadow p-6">
+          <h2 className="text-lg font-semibold mb-4">Cache Performance</h2>
+          <CacheStatus />
         </div>
       )}
     </DashboardLayout>
