+++ src/frontend/nextjs-app/app/types/demand-planning/forecast.ts
@@ -2,7 +2,6 @@
  * Types for forecast data in the demand planning dashboard
  */
 
-import { HierarchySelection } from './hierarchy';
 
 // Time period definition
 export interface TimePeriod {
@@ -37,7 +36,6 @@ export interface InventoryItem {
 // Complete forecast series
 export interface ForecastSeries {
   id: string;
-  hierarchySelections: HierarchySelection[];
   timePeriods: TimePeriod[];
   baseline: ForecastDataPoint[];
   adjusted?: ForecastDataPoint[];
@@ -62,7 +60,6 @@ export type AdjustmentReason =
 // Adjustment definition
 export interface Adjustment {
   id: string;
-  hierarchySelections: HierarchySelection[];
   timePeriods: string[]; // Array of period IDs
   type: AdjustmentType;
   value: number;
